+++ tests/getServiceProxy.test.js
@@ -5,7 +5,15 @@ const nock = require('nock')
 
 const MY_AWESOME_SERVICE_PROXY_URL = 'my-awesome-service'
 const MICROSERVICE_GATEWAY_SERVICE_NAME = 'microservice-gateway'
-const { getDirectServiceProxy, getServiceProxy } = require('../index')
+const MY_AWESOME_SERVICE_PROXY_HTTP_URL = 'http://my-awesome-service'
+const MY_AWESOME_SERVICE_PROXY_HTTPS_URL = 'https://my-awesome-service'
+const MY_AWESOME_SERVICE_PROXY_HTTP_URL_CUSTOM_PORT = 'http://my-awesome-service:3000'
+const MY_AWESOME_SERVICE_PROXY_HTTPS_URL_CUSTOM_PORT = 'https://my-awesome-service:3001'
+
+const {
+  getDirectServiceProxy,
+  getServiceProxy,
+} = require('../index')
 
 tap.test('getDirectServiceProxy available for testing', async t => {
   nock.disableNetConnect()
@@ -72,4 +80,103 @@ tap.test('getServiceProxy available for testing', async t => {
   t.strictSame(result.statusCode, 200)
   t.strictSame(result.payload.message, RETURN_MESSAGE)
   microserviceScope.done()
+})
+
+tap.test('getDirectServiceProxy available for testing - complete url passed', async t => {
+  nock.disableNetConnect()
+  t.tearDown(() => {
+    nock.enableNetConnect()
+  })
+
+  const RETURN_MESSAGE = 'OK'
+  const customProxy = getDirectServiceProxy(MY_AWESOME_SERVICE_PROXY_HTTP_URL)
+  const awesomeHttpServiceScope = nock(`${MY_AWESOME_SERVICE_PROXY_HTTP_URL}:80`)
+    .get('/test-endpoint')
+    .reply(200, {
+      message: RETURN_MESSAGE,
+    })
+
+  const result = await customProxy.get('/test-endpoint')
+
+  t.strictSame(result.statusCode, 200)
+  t.strictSame(result.payload.message, RETURN_MESSAGE)
+  awesomeHttpServiceScope.done()
+})
+
+tap.test('getDirectServiceProxy available for testing - https url', async t => {
+  nock.disableNetConnect()
+  t.tearDown(() => {
+    nock.enableNetConnect()
+  })
+
+  const RETURN_MESSAGE = 'OK'
+  const customProxy = getDirectServiceProxy(MY_AWESOME_SERVICE_PROXY_HTTPS_URL)
+  const awesomeHttpsServiceScope = nock(`${MY_AWESOME_SERVICE_PROXY_HTTPS_URL}:443`)
+    .get('/test-endpoint')
+    .reply(200, {
+      message: RETURN_MESSAGE,
+    })
+
+  const result = await customProxy.get('/test-endpoint')
+
+  t.strictSame(result.statusCode, 200)
+  t.strictSame(result.payload.message, RETURN_MESSAGE)
+  awesomeHttpsServiceScope.done()
+})
+
+tap.test('getDirectServiceProxy available for testing - custom port 3000 - custom headers', async t => {
+  nock.disableNetConnect()
+  t.tearDown(() => {
+    nock.enableNetConnect()
+  })
+
+  const RETURN_MESSAGE = 'OK'
+  const customProxy = getDirectServiceProxy(MY_AWESOME_SERVICE_PROXY_HTTP_URL_CUSTOM_PORT,
+    {
+      headers: {
+        'test-header': 'test header works',
+      },
+    })
+  const awesomeHttpServiceScope = nock(`${MY_AWESOME_SERVICE_PROXY_HTTP_URL}:3000`)
+    .matchHeader('test-header', 'test header works')
+    .get('/test-endpoint')
+    .reply(200, {
+      message: RETURN_MESSAGE,
+    })
+
+  const result = await customProxy.get('/test-endpoint')
+
+  t.strictSame(result.statusCode, 200)
+  t.strictSame(result.payload.message, RETURN_MESSAGE)
+  awesomeHttpServiceScope.done()
+})
+
+tap.test('getDirectServiceProxy available for testing - https url - custom port 3001', async t => {
+  nock.disableNetConnect()
+  t.tearDown(() => {
+    nock.enableNetConnect()
+  })
+
+  const RETURN_MESSAGE = 'OK'
+  const customProxy = getDirectServiceProxy(MY_AWESOME_SERVICE_PROXY_HTTPS_URL_CUSTOM_PORT)
+  const awesomeHttpsServiceScope = nock(`${MY_AWESOME_SERVICE_PROXY_HTTPS_URL}:3001`)
+    .get('/test-endpoint')
+    .reply(200, {
+      message: RETURN_MESSAGE,
+    })
+
+  const result = await customProxy.get('/test-endpoint')
+
+  t.strictSame(result.statusCode, 200)
+  t.strictSame(result.payload.message, RETURN_MESSAGE)
+  awesomeHttpsServiceScope.done()
+})
+
+tap.test('getDirectServiceProxy throws on invalid url', async t => {
+  const invalidUrl = 'httpnot-a-complete-url'
+  try {
+    getDirectServiceProxy(invalidUrl)
+  } catch (error) {
+    t.notOk(true, 'The function should not throw anymore if the url is not a valid one, bet return the standard proxy')
+  }
 })